🐛 Fix: type 수정

+++ backend/franchise/src/main/java/kr/co/olivepay/franchise/dto/res/PagedFranchiseReviewsRes.java
@@ -5,11 +5,13 @@ import java.util.List;
 import lombok.Builder;
 
 /**
- *
+ * 내가 작성한 리뷰와 페이징 처리를 위한 index를 담은 dto
+ * @param nextIndex
+ * @param reviews
  */
 @Builder
 public record PagedFranchiseReviewsRes(
-	int nextIndex,
+	Long nextIndex,
 	List<FranchiseReviewRes> reviews
 ) {
 }
